refactor(org2cloud): split org2CloudSyncClient into rpc/wire siblings
Move the RPC plumbing, zod wire contract, org listing, session-events
plane, org-governance RPCs and the 0012 turn index into dotted siblings;
org2CloudSyncClient.ts is now a pure re-export barrel so all 21 import
sites are unchanged. Each capability cache Set stays in the one module
holding both its consumers and its __*_INTERNALS reset hook. No
behaviour change.
